requirements.txt
google-generativeai>=0.3.0
langchain-google-genai>=2.0.10
langchain-core>=0.3.50
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
tenacity>=8.2.3
gradio>=4.0.0
cartesia>=2.0.0
matplotlib>=3.0.0
faiss-cpu>=1.10.0
numpy>=1.24.0
pillow>=8.0.0